content-security-policy/script-src/eval-allowed-in-report-only-mode-and-sends-report.html

<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <!-- Content-Security-Policy-Report-Only: script-src 'unsafe-inline'; report-uri /reporting/resources/report.py?op=put&reportID={{$id}} -->
</head>
<body>
  <script>
    eval = new Proxy(eval, {
            apply:
                function(target, thisArg, args) {
                  try{
                    // console.log("[{{host}}]: [eval] -> " + args );
                    let res = target.apply(thisArg, args);
                    console.log("success");
                    return res;
                  } catch (e){
                    console.log("fail");
                    throw e;
                  }
                }
        });
    var t = async_test("Eval is allowed because the CSP is report-only");
    try {
      eval("t.done()");
    } catch (e){
      console.log(e);
      t.step(function() { assert_true(false, "The eval should have execute succesfully"); })
    }
  </script>

  <script async defer src="../support/checkReport.sub.js?reportField=blocked-uri&reportValue=eval"></script>
</body>
</html>
